requirements.txt

requests
beautifulsoup4
selenium
webdriver-manager
pandas
streamlit
python-dotenv
tqdm
